Passes full set of oauth consumer parameters
The LinkedIn::Helpers::Authorization.request_token function did not pass the full set of parameters to the OAuth::Consumer's request_token function even though it was just a pass through.

This was causing issues with requesting additional scope parameters from LinkedIn.

This commit adds a new spec to test that things still work with the additional parameters,
but it does not contain a new VCR cassette since I don't have the keys used to generate
the existing requests.  This needs to be added.

I'd like to also test the the scope is actually sent, but I don't currently have the time
to lookup the required functions in the HTTP object.