Add ability to send targeted company shares
This introduces a breaking change to the add_company_share method. We
now parse the response body so the user does not have to parse the xml
response body. This change should not impact many users seeing as the
add_company_share method was just added a while ago and there hasn't
been a release since.

Usage:

    response = client.add_company_share("2414183", {
      :comment => "Testing, 1, 2, 3",
      :targets => {
        :geos => ['as', 'eu'],
        :jobFunc => ['acct', 'bd']
      }
    puts "ID: #{response.body.update_key}"
    puts "URL: #{response.body.update_url}"